requirements.txt
Flask==2.1.1
Flask-Jsonpify==1.5.0
requests==2.18.4
requests-unixsocket==0.1.5
Werkzeug==2.0.1
PyYAML==3.12